+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-
+### Added
+- Parameters do optionally accept a factory function to create the instance.
 
 ## [0.6.1] - 2016-05-17
 ### Fixed