+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+
 ## [Unreleased]
+### Added
+- `Query` and `Body` parameters do optionally accept a factory function to create the instance.
+
+### Changed
+- Updated `typings` to 1.x version: `ambient` dependencies are new `global` dependencies.
 
 
 ## [0.6.1] - 2016-05-17